rebranding from debug to ci testing
