+++ crates/bevy_internal/Cargo.toml
@@ -41,8 +41,8 @@ x11 = ["bevy_winit/x11"]
 # enable rendering of font glyphs using subpixel accuracy
 subpixel_glyph_atlas = ["bevy_text/subpixel_glyph_atlas"]
 
-# enable debug systems, for Bevy development
-bevy_debug = ["bevy_app/bevy_debug"]
+# enable systems that allow for automated testing on CI
+bevy_ci_testing = ["bevy_app/bevy_ci_testing"]
 
 [dependencies]
 # bevy
